Fix PEP8 violations in Sheet() and remove getopt
The doc string of the Sheet() class went past column 80 in several
places. The display of the various Excel worksheet layouts (key-value,
list, dictionary) was off due to bad formatting in the text editor that
the file was originally built in.

The getopt module was no longer needed after it was replaced with
argparse for command line option gathering, but it was still being
imported.