Add support for table-formatted worksheets
Previously, xlyaml only supported object parsing in list format. List
format is good when the objects in a sheet have varying numbers of
attributes or need to support list or dictionary type structures.
However, for a fixed number of key-value variable types across objects,
a simple table format can be easier to document in Excel. xlyaml now
supports this formatting.